Add ahead/behind determination to git prompt
Uses git rev-parse to determine how far ahead or behind we are
from the upstream branch, with + denoting # of commits that do not
appear on the upstream, and - denoting # commits that do not appear
on the local branch.

If the branch does not have an upstream, or if the repository is on
its initial commit, the ahead/behind determination works as desired.